healthcheck/requirements.txt
PyYAML == 6.0
jsonschema == 4.17.3
pymongo == 3.12
kink == 0.6.6
boto3 == 1.26.91
boto3-stubs[s3,sqs] == 1.26.91
botocore == 1.29.91
aws_error_utils == 2.7.0
expiringdict == 1.2.2
--extra-index-url https://${TOKEN}@pypi.fiftyone.ai
-e ../base[aws,voxel,model]
